cleaned up classes and fixed some broken movement code
